pyproject.toml
[project]
name = "weatherman"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "boto3>=1.37.18",
    "fastapi[standard]>=0.115.11",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "opencage>=3.0.4",
    "pvlib>=0.12.0",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
]
